Create RESULTS.MD
This first benchmark result was made "manually", by running it multiple times on simulator and pasting the results on this file.
TODO: Create automation to run all benchmarks every x period of time and update this file.
Travis CI and Detox tool from wix will probably help with that.

Environment:
  MacBook Pro 16GB 2.9 GHz Intel Core i5
  OS: macOS High Sierra 10.13.1
  Xcode: Xcode 9.1 Build version 9B37
  Device: iPhone 8 Simulator iOS 11.1
  Build: Release mode